patch #1661,1662,1675,1680-1685,1687-1691,1693,1700,1702,1705-1707 this is completely untested I will start two builds now and check in fixes as I find them Robin, you can start builds as well, if you want to

+++ programs/Xserver/hw/xfree86/accel/glint/glint_accel.c
@@ -1,4 +1,4 @@
-/* $XFree86: xc/programs/Xserver/hw/xfree86/accel/glint/glint_accel.c,v 1.22 1998/04/05 00:45:48 robin Exp $ */
+/* $XFree86: xc/programs/Xserver/hw/xfree86/accel/glint/glint_accel.c,v 1.23 1998/04/26 16:04:38 robin Exp $ */
 /*
  * Copyright 1996,1997 by Alan Hourihane, Wigan, England.
  *
@@ -722,9 +722,9 @@ GLINTWriteBitmap(x, y, w, h, src, srcwidth, srcx, srcy,
 
     srcp = (srcwidth * srcy) + (srcx >> 3) + src; 
     srcx &= 0x07;
-    if(skipleft = (int)srcp & 0x03) { 
+    if(skipleft = (long)srcp & 0x03) { 
 	skipleft = (skipleft << 3) + srcx;	
-	srcp = (unsigned char *)((int)srcp & ~0x03);
+	srcp = (unsigned char *)((long)srcp & ~0x03);
     } else
 	skipleft = srcx;
